Update design and code

+++ app/src/main/java/ru/floppastar/shopcashier/db/dbRepository.kt
@@ -1,8 +1,10 @@
 package ru.floppastar.shopcashier.db
 
 import android.content.ContentValues
+import android.content.Context
 import android.database.Cursor
 import android.provider.ContactsContract.Data
+import android.widget.Toast
 import ru.floppastar.shopcashier.DataClasses.GoodsInStock
 import ru.floppastar.shopcashier.DataClasses.GoodsType
 import java.text.SimpleDateFormat
@@ -10,7 +12,6 @@ import java.util.Date
 import java.util.Locale
 
 class dbRepository(private val dbHelper: DatabaseHelper) {
-    //Типы товаров
     fun insertGoodsType(name: String) {
         val db = dbHelper.writableDatabase
         val values = ContentValues().apply {
@@ -51,16 +52,29 @@ class dbRepository(private val dbHelper: DatabaseHelper) {
         db.close()
     }
 
-    fun deleteGoodsType(goodsTypeId: Int) {
-        val db = dbHelper.writableDatabase
+    fun deleteGoodsType(goodsTypeId: Int, context: Context): Boolean {
+        var db = dbHelper.readableDatabase
+        val query = "SELECT COUNT(*) FROM ${DatabaseHelper.TABLE_GOODS_IN_STOCK} WHERE ${DatabaseHelper.COLUMN_GOODS_IN_STOCK_TYPE} = ?"
+        val cursor = db.rawQuery(query, arrayOf(goodsTypeId.toString()))
+        var count = 0
+        if (cursor.moveToFirst()) {
+            count = cursor.getInt(0)
+        }
+        cursor.close()
+        db.close()
+        if (count > 0){
+            Toast.makeText(context, "Невозможно удалить тип товара, так как он используется в одном или нескольких товарах", Toast.LENGTH_SHORT).show()
+            return false
+        }
+        db = dbHelper.writableDatabase
         db.delete(
             DatabaseHelper.TABLE_GOODS_TYPE,
             "${DatabaseHelper.COLUMN_GOODS_TYPE_ID} = $goodsTypeId", null
         )
         db.close()
+        return true
     }
 
-    //Товары на складе
     fun insertGoodsInStock(name: String, goodsCount: Int, price: Double, goodsTypedId: Int) {
         val db = dbHelper.writableDatabase
         val values = ContentValues().apply {
@@ -113,6 +127,10 @@ class dbRepository(private val dbHelper: DatabaseHelper) {
     fun deleteGoodsInStock(goodId: Int) {
         val db = dbHelper.writableDatabase
         db.delete(
+            DatabaseHelper.TABLE_GOODS_SOLD,
+            "${DatabaseHelper.COLUMN_GOODS_SOLD_GOOD} = $goodId", null
+        )
+        db.delete(
             DatabaseHelper.TABLE_GOODS_IN_STOCK,
             "${DatabaseHelper.COLUMN_GOODS_IN_STOCK_ID} = $goodId", null
         )
@@ -122,14 +140,12 @@ class dbRepository(private val dbHelper: DatabaseHelper) {
     fun getGoodsTypeName(goodsTypeId: Int): String {
         val db = dbHelper.readableDatabase
         var goodType = GoodsType(1,"")
-
         var cursor: Cursor? = null
         try {
             cursor = db.rawQuery(
                 "SELECT ${DatabaseHelper.COLUMN_GOODS_TYPE_NAME} FROM ${DatabaseHelper.TABLE_GOODS_TYPE} " +
                         "WHERE ${DatabaseHelper.COLUMN_GOODS_TYPE_ID} = ?", arrayOf(goodsTypeId.toString())
             )
-
             if (cursor != null && cursor.moveToFirst()) {
                 goodType.name = cursor.getString(cursor.getColumnIndexOrThrow(DatabaseHelper.COLUMN_GOODS_TYPE_NAME))
             }
@@ -137,11 +153,9 @@ class dbRepository(private val dbHelper: DatabaseHelper) {
             cursor?.close()
             db.close()
         }
-
         return goodType.name
     }
 
-    // Получение или создание saleId для текущей продажи
     fun getOrCreateSaleId(): Int {
         val db = dbHelper.readableDatabase
         val cursor = db.rawQuery(
@@ -162,7 +176,6 @@ class dbRepository(private val dbHelper: DatabaseHelper) {
         }
     }
 
-    // Добавление товара в корзину
     fun addGoodsToCart(goodId: Int, saleId: Int): Boolean {
         val db = dbHelper.writableDatabase
         db.beginTransaction()
@@ -176,7 +189,6 @@ class dbRepository(private val dbHelper: DatabaseHelper) {
                 stockCount = cursor.getInt(0)
             }
             cursor.close()
-
             val goodsInCartCount = getGoodsCountInCart(goodId, saleId)
             if (goodsInCartCount >= stockCount) {
                 return false
@@ -193,19 +205,16 @@ class dbRepository(private val dbHelper: DatabaseHelper) {
         }
     }
 
-    // Удаление товара из корзины
     fun removeGoodsFromCart(goodId: Int, saleId: Int) {
         val db = dbHelper.writableDatabase
         db.beginTransaction()
         try {
-            // Получить ID первой записи, которая соответствует критериям
             val cursor = db.rawQuery(
                 "SELECT ${DatabaseHelper.COLUMN_GOODS_SOLD_ID} FROM ${DatabaseHelper.TABLE_GOODS_SOLD} WHERE ${DatabaseHelper.COLUMN_GOODS_SOLD_GOOD} = ? AND ${DatabaseHelper.COLUMN_GOODS_SOLD_SALE} = ? LIMIT 1",
                 arrayOf(goodId.toString(), saleId.toString())
             )
             if (cursor.moveToFirst()) {
                 val id = cursor.getInt(0)
-                // Удалить запись по ID
                 db.execSQL("DELETE FROM ${DatabaseHelper.TABLE_GOODS_SOLD} WHERE ${DatabaseHelper.COLUMN_GOODS_SOLD_ID} = ?", arrayOf(id.toString()))
             }
             cursor.close()
@@ -215,7 +224,6 @@ class dbRepository(private val dbHelper: DatabaseHelper) {
         }
     }
 
-    // Получение количества добавленных товаров
     fun getGoodsCountInCart(goodId: Int, saleId: Int): Int {
         val db = dbHelper.readableDatabase
         val cursor = db.rawQuery("SELECT COUNT(*) FROM ${DatabaseHelper.TABLE_GOODS_SOLD} WHERE ${DatabaseHelper.COLUMN_GOODS_SOLD_GOOD} = ? AND ${DatabaseHelper.COLUMN_GOODS_SOLD_SALE} = ?", arrayOf(goodId.toString(), saleId.toString()))
@@ -230,7 +238,6 @@ class dbRepository(private val dbHelper: DatabaseHelper) {
         val db = dbHelper.writableDatabase
         db.beginTransaction()
         try {
-            // 1. Обновление количества товаров на складе
             val cursor = db.rawQuery("SELECT ${DatabaseHelper.COLUMN_GOODS_SOLD_GOOD}, COUNT(${DatabaseHelper.COLUMN_GOODS_SOLD_GOOD}) AS count FROM ${DatabaseHelper.TABLE_GOODS_SOLD} WHERE ${DatabaseHelper.COLUMN_GOODS_SOLD_SALE} = ? GROUP BY ${DatabaseHelper.COLUMN_GOODS_SOLD_GOOD}", arrayOf(saleId.toString()))
             while (cursor.moveToNext()) {
                 val goodId = cursor.getInt(cursor.getColumnIndexOrThrow(DatabaseHelper.COLUMN_GOODS_SOLD_GOOD))
@@ -248,20 +255,15 @@ class dbRepository(private val dbHelper: DatabaseHelper) {
                 stockCursor.close()
             }
             cursor.close()
-
-            // 2. Обновление записи SaleInformation
             val contentValues = ContentValues().apply {
                 put(DatabaseHelper.COLUMN_SALE_INFORMATION_PURCHASE_DATE, getCurrentDate())
             }
             db.update(DatabaseHelper.TABLE_SALE_INFORMATION, contentValues, "${DatabaseHelper.COLUMN_SALE_INFORMATION_ID} = ?", arrayOf(saleId.toString()))
-
-            // 3. Создание новой записи SaleInformation
             val newSaleContentValues = ContentValues().apply {
                 put(DatabaseHelper.COLUMN_SALE_INFORMATION_PURCHASE_DATE, "")
                 put(DatabaseHelper.COLUMN_SALE_INFORMATION_TOTAL_PRICE, 0)
             }
             db.insert(DatabaseHelper.TABLE_SALE_INFORMATION, null, newSaleContentValues)
-
             db.setTransactionSuccessful()
             return true
         } finally {
